requirements.txt
asgiref==3.7.2 
django>=2.2,<5.0
sqlparse==0.4.4 
tzdata==2023.3
djangorestframework==3.14.0 
pytz==2023.3.post1
django-phonenumber-field==7.1.0 
phonenumbers==8.13.21
djangorestframework-simplejwt==5.3.0 
pyjwt==2.8.0
django-countries==7.5.1 
typing-extensions==4.8.0
django-randompinfield==0.4.7
certifi==2023.7.22 
charset-normalizer==3.2.0 
django-twilio==0.14.3.2 
idna==3.4 
requests==2.31.0 
twilio==7.17.0 
urllib3==2.0.5
Pillow==10.0.1